Refactored Evaluator class to return a float for evaluateExpression method

+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/actions/draw/DrawCardsAction.java
@@ -51,7 +51,7 @@ public class DrawCardsAction extends ActionyAction implements TopLevelSelectable
     public boolean requirementsAreMet(DefaultGame cardGame) {
         try {
             Player performingPlayer = cardGame.getPlayer(_performingPlayerId);
-            return performingPlayer.getCardsInDrawDeck().size() >= _cardDrawCountEvaluator.evaluateExpression(cardGame);
+            return performingPlayer.getCardsInDrawDeck().size() >= (int) _cardDrawCountEvaluator.evaluateExpression(cardGame);
         } catch(PlayerNotFoundException exp) {
             cardGame.sendErrorMessage(exp);
             return false;
@@ -60,7 +60,7 @@ public class DrawCardsAction extends ActionyAction implements TopLevelSelectable
 
     @Override
     public Action nextAction(DefaultGame cardGame) throws InvalidGameLogicException, PlayerNotFoundException {
-        int totalDrawCount = _cardDrawCountEvaluator.evaluateExpression(cardGame);
+        int totalDrawCount = (int) _cardDrawCountEvaluator.evaluateExpression(cardGame);
         if (isBeingInitiated())
             setAsInitiated();
         if (_cardsAlreadyDrawnCount < totalDrawCount) {
