+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/evaluator/SingleMemoryEvaluator.java
@@ -1,11 +1,9 @@
 package com.gempukku.stccg.evaluator;
 
-import com.gempukku.stccg.cards.ActionContext;
-import com.gempukku.stccg.cards.physicalcard.PhysicalCard;
 import com.gempukku.stccg.game.DefaultGame;
 
 public class SingleMemoryEvaluator extends Evaluator {
-    private Integer _rememberedValue;
+    private Float _rememberedValue;
     private final Evaluator _evaluator;
 
     public SingleMemoryEvaluator(Evaluator evaluator) {
@@ -14,7 +12,7 @@ public class SingleMemoryEvaluator extends Evaluator {
     }
 
     @Override
-    public int evaluateExpression(DefaultGame game) {
+    public float evaluateExpression(DefaultGame game) {
         if (_rememberedValue == null)
             _rememberedValue = _evaluator.evaluateExpression(game);
         return _rememberedValue;
